Refactor. Moved code outside of the hardware drivers that doesn't belong there.
