+++ include/hardware/lgy.h
@@ -103,14 +103,12 @@ typedef struct
 
 
 
+Result LGY_prepareGbaMode(bool biosIntro, u16 saveType, const char *const savePath);
 Result LGY_setGbaRtc(const GbaRtc rtc);
 Result LGY_getGbaRtc(GbaRtc *const out);
 Result LGY_backupGbaSave(void);
 #ifdef ARM11
-Result LGY_prepareGbaMode(bool biosIntro, char *const romPath);
 void LGY_switchMode(void);
-void LGY_handleEvents(void);
+void LGY_handleOverrides(void);
 void LGY_deinit(void);
-#elif ARM9
-Result LGY_prepareGbaMode(bool biosIntro, u16 saveType, const char *const savePath);
 #endif
